# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kontra orchestrator",
    "version": "0.1.0",
    "description": "Generated from the routes this server registers. Paths and auth are authoritative; request and response bodies are deliberately absent until routes carry JSON schemas."
  },
  "components": {
    "securitySchemes": {
      "consoleSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "A session from POST /api/login."
      },
      "stateToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "KONTRA_STATE_TOKEN. Can spend money."
      },
      "runToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "KONTRA_RUN_TOKEN. Open when unset."
      },
      "exploreToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "KONTRA_EXPLORE_TOKEN, falling back to state."
      },
      "actorIdentity": {
        "type": "http",
        "scheme": "bearer",
        "description": "A signed token naming ONE actor, minted when it is served (KONTRA_ACTOR_TOKEN). It resolves what that actor owns and nothing more — it is not a service token."
      }
    }
  },
  "paths": {
...
    "/api/infra/ops/{fqn}": {
      "get": {
        "operationId": "getInfraOpsByFqn",
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "security": [
          {
            "stateToken": []
          },
          {
            "consoleSession": []
          }
        ],
        "description": "Gated by KONTRA_STATE_TOKEN, or a console session (ADR 0045)."
      }
    },
    "/api/infra/stacks": {
      "get": {
        "operationId": "getInfraStacks",
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "security": [
          {
            "stateToken": []
          },
          {
            "consoleSession": []
          }
        ],
        "description": "Gated by KONTRA_STATE_TOKEN, or a console session (ADR 0045)."
      }
    },
    "/api/infra/stacks/{fqn}/{op}": {
      "post": {
        "operationId": "postInfraStacksByFqnByOp",
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "security": [
          {
            "stateToken": []
          },
          {
            "consoleSession": []
          }
        ],
        "description": "Gated by KONTRA_STATE_TOKEN, or a console session (ADR 0045)."
      }
    },
    "/api/infra/stacks/{fqn}/leases": {
      "get": {
        "operationId": "getInfraStacksByFqnLeases",
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "security": [
          {
            "stateToken": []
          },
          {
            "consoleSession": []
          }
        ],
        "description": "Gated by KONTRA_STATE_TOKEN, or a console session (ADR 0045)."
      }
    },
    "/api/infra/stacks/{fqn}/state": {
      "get": {
        "operationId": "getInfraStacksByFqnState",
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "security": [
          {
            "stateToken": []
          },
          {
            "consoleSession": []
          }
        ],
        "description": "Gated by KONTRA_STATE_TOKEN, or a console session (ADR 0045)."
      }
    },
...
  }
}